PtyProcess: add NO_CTTY flag
The default behavior of posix_openpt seems to be to replace the
controlling terminal for the calling process, and PtyProcess should
only change it for child instead. I think this might be a reason why
some of the tests were failing non-deterministacally sometimes, although
I am not 100% confident in this fix.

Signed-off-by: Petre Eftime <petre.eftime@gmail.com>